change input date format
Without having made any changes on our side, Flight Schedule Pro appears
to have changed the date format they use in reservation notification
emails. The format is now %d/%m/%Y .

Here is an excerpt from a reservation notification email received on
2013-12-30 for a reservation the same day:

Start Time: 12/30/2013 7:00 PM

This format cannot be interpreted as anything but %m/%d/%Y (an
unfortunate choice of format).

Here is an excerpt from a reservation notification email received on
2014-01-13, also for a reservation the same day:

Start Time: 13/01/2014 4:30 PM

This format cannot be interpreted as anything but %d/%m/%Y .

The locale settings on the account were reported to have not been
changed. We have to assume something changed on the Flight Schedule
Pro end and that dates will be henceforth in %d/%m/%Y format.